Int=>Double(How Many)

+++ ExpensesAPI/Dto/PurchaseDto.cs
@@ -7,7 +7,7 @@
         [MaxLength(50)]
         public string Description { get; set; } = string.Empty;
         [Required]
-        public int HowMany { get; set; }
+        public double HowMany { get; set; }
         public string Сurrency { get; set; } = string.Empty;
         public DateTime WasBought { get; set; }
         [Required]
